Service loading for the datastore client
Summary:
This commit implements a `ServiceLoader` for a `DatastoreBackend` interface,
which will wrap the google cloud client library (or test stub) with environment
implemenation, allowing our eventual `DB` singleton object to choose a backend
based on whether we're in the prod, dev, or test environment. We'll eventually
provide two implemenations: a prod/dev one using the google cloud client
library (using its internal functionality to use the emulator as needed in
dev), and a test stub. (This pattern also has the side effect that clients of
this library could provide their own implementation if desired.)

There's a couple complications here that make this more complex than a naive
implementation might be:

1. I wanted to be able to assert that we only provide at most one
  implementation per environment (and provide a good error message when this is
  violated). In the past, I've found it easy, especially in multi-project
  builds, to accidentally load service implementations I didn't intend
  to. Because the service is going to be consumed internal to this library (and
  we have no general purpose way to tie break between implementations), we need
  to handle this case consistently, which I've done by throwing.

2. `ServiceLoader` is a final class that is nontrivial to use with test stubs,
  so referring to it directly makes it hard to test the loading functionality
  itself. I therefore wrapped ServiceLoader in another layer that we can use to
  provide a mock loader for testing, which allows us to test the environment
  selection functionality. This is a bit ugly, but it doesn't contaminate the
  public API of this library, so I think it's probably ok in the name of better
  testability.

In a follow-up commit, I'll scaffold the `DB` singleton object that is the
primary public API of this subproject as well as the consumer of this service
loader.

Test Plan: - `./gradlew test`

Reviewers: csilvers, aasmund

Reviewed By: csilvers, aasmund

Differential Revision: https://phabricator.khanacademy.org/D52000