Implement transactions + async operations
Summary:
I think the trickiest thing we have to do in this project is to make sure that
transactions work correctly when combined with asynchronous operations, so
let's implement that first, to make sure we can actually achieve the API we
want!

The reason this is tricky is the particulars of how async code executes in
kotlin. For our purposes, we're running the coroutines in a background thread
pool, so for the most part, they operate indistinguishably from lightweight
threads, rather than traditional coroutines in the style of python asyncio, or
ndb tasklets, or javascript async/await. The one difference is that a kotlin
coroutine may execute in many threads in a thread pool over its lifetime, so
java thread-local variables don't work as expected, and we need to do some
extra work to make them coroutine-local variables instead.

The goals I wanted to acheive are:
- all datastore access goes through the singleton `DB` object
- transactionality follows how the code reads, not how the code executes. That
  is, if you have a block of `DB.transactional { }` code, then everything in that
  block participates in that transaction, regardlesss of how many coroutines or
  threads it may run over.

The implementation is a bit complicated, but hopefully the good complicated,
where we're insulating clients of the library from the complications. (We'll
have to keep an eye on usage and make sure that the abstractions here are not
leakier than I think they are.) All datastore operations are defined on a
`DatastoreContext`, which either tracks nothing special if we're outside a
transaction, or the ongoing transaction. These also implement `CoroutineScope`,
which we exploit to make the context track with coroutine execution.

`DB` is then a thread local that's made into a coroutine local by including it
in the coroutine context. Kotlin provides a helper function for doing
this (`ThreadLocal.asContextElement()`), which makes the runtime take care of
saving the value of the thread local in the coroutine context, setting up that
value when a coroutine begins executing in a thread, and then tearing it down
when the coroutine suspends. This allows us to have the global `DB` singleton
still be context-aware enough to put operations in a transaction when
necessary.

Test Plan: - `./gradlew test`

Reviewers: csilvers, aasmund, benkraft

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D52088