Implement reflective entity conversion to data classes
Summary:
We're planning to represent datastore models by plain (possibly annotated) data
classes in kotlin. We therefore need to convert these back and forth to the
`Entity` format used by the datstore client, which is very similar to a
`Map<String, Any?>`. This commit implements the conversion in one direction,
from `Entity` to data class, using kotlin's reflection capabilities.

Mostly this was straightforward, but there's a couple design choices worth
noting:
1. The java client library distinguishes between the cases where a property is
  not set and when the property is null. I collapsed this into just the notion
  of nullability on the data classes. It wasn't clear to me what it would mean
  for a property to be optional separate from nullable on a data class, and in
  any case, I didn't want to further the `null` vs. `undefined` confusion we
  know and love in javascript.

2. Any mismatch between the stored entity and the data class's schema throws an
  exception when we're trying to do the conversion. There was also the
  possibility that we could just have the `.get()` return null in this case,
  but I think that a) it's better to reserve the null value to mean the key is
  not in the datastore b) since a mismatch here reflects a problem with the
  code, rather than a possibly expected condition, it's better to fail quickly.

As a practical matter, I also found that I needed a reference to the class in
order to do this conversion, and since the `get()` function is defined as part
of the `DatastoreContext` interface, it couldn't use a reified type parameter,
so I changed the interface for `.get()` to have an explicit class reference
passed in, and I added top-level convenience functions that don't require this
additional reference, which I expect will be the normal way to do `.get()`s.

Test Plan: - `./gradlew test`

Reviewers: csilvers, aasmund, benkraft

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D52156