Add an annotation for property renaming
Summary:
In many cases, to keep in line with style conventions, we'll want our datastore
model properties to have a different (camel-cased) name in code from
their (underscore-cased, due to python) name in the datastore. Additionally,
there are a few miscellaneous cases where the name in the datastore is
different even from its name in python that we'll need to be able to handle.

This commit adds a runtime-available annotation called `Meta` that we'll use to
annotate properties with datastore-specific metadata (like name, and whether a
property is indexed or not). Then, I've added a function to pull out the name
based on this annotation when we're converting `Entity`s to our model data
classes.

Test Plan: - `./gradlew test`

Reviewers: csilvers, aasmund, benkraft

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D52187