Set up API documentation generation
Summary:
In order to pull the datastore library into webapp, I'm planning to publish it
as a package to the maven central repository (as a prerelease version). There's
some automation to do as part of this, and one thing we'd like is to have
public API documentation available alongside the jar file as a reference.

The kotlin documentation tool is called `dokka`; this commit sets us up to use
it to generate combined documentation for both subprojects.

We're not doing anything with it yet; that will come in a later commit with
some jenkins automation. The documentation is also pretty sparse at this point,
but this will at least provide us a foundation on which to fill it out in the
expected format for comments.

Test Plan:
- `./gradlew dokka`
- go to build/docs
- `python3 -m http.server`
- visit localhost:8000
- click around, see that there's documenation of the public parts of both
  subprojects.

Reviewers: csilvers, aasmund, benkraft

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D52248