Update the README to reflect high-level project status and add a documentation link
Summary:

I've had folks asking about the status of the kotlin datastore library project,
so let's make sure that there's a place we can point them to that shows an
overview of what features are and are not implemented. (This is also just a
nice thing to have in general since this is a public project.)

I also pushed the generated documenation (via `./gradlew dokka`) to the
gh-pages branch of this repository and added a link to the readme. It's pretty
sparse at the moment (I should be more careful about commenting in a way that
the tool picks up), but it's a good start.

Test Plan:
- fingers crossed

Auditors: csilvers, aasmund, benkraft