Add a partial simple test stub for clients of the library
Summary:
In order for clients to be able to run tests using this library, we need a
test-mode stub. This commit is a start on an implementation that has the
following behavior:
- we by default throw on all datastore operations (to ensure that you must
  specify behavior explicitly for all tests that access the datastore)
- you can within a block of code provide a lightweight mock for the datastore,
  which contains a list of entities that you provide at the start of the block

At the moment this is pretty incomplete, and the only thing that's supported is
get-by-key, but this is enough of a start to allow us to pull it into webapp.

Additionally at some point, we should add a second option to use the datastore
emulator as a backend. This is considerably more heavyweight, but it'll provide
a more accurate simulation of the datastore. The lightweight stub this commit
adds is intended for more casual use, where in python we might, say, have
mocked out a getter function to return a specfic value (rather than using the
full-on test DB).

Test Plan: - `./gradlew test`

Reviewers: csilvers, aasmund, benkraft

Reviewed By: csilvers

Subscribers: marksandstrom

Differential Revision: https://phabricator.khanacademy.org/D52317