Set up publishing of build artifacts to GCS
Summary:
In order to be able to consume this library from webapp (or eventually for
anyone else), we need to publish the built code to a maven repository.

There's a standard one, maven central, to which we could publish, but after a
bit of poking around and reading up, I gave up temporarily (my eyes started to
glaze over at the point where I needed to generate GPG signing keys). It's
pretty easy just to use GCS as a maven repository, so this commit implements
that. We can figure out how to publish to maven central later, or (less
optimally) just make our GCS bucket public.

Note that I've set this up so that we publish the two subprojects
separately. This is intentional because we'd like in some cases (e.g. for
bigquery-related code) to be able to pull in the schema data classes, which
depend on the metadata annotations, without pulling in the dependency on the
datastore client library.

There's still a lot here to do... we'll want to set up a jenkins job to do this
instead of doing it locally, we'll also probably want to automate bumping the
version number so we don't accidentally release the same version twice,
etc. But for now, this gets us off the ground.

Test Plan:
- `./gradlew publish`
- verify I see jarfiles in our GCS bucket
- verify I can pull them in as a dependency in webapp (commit incoming soon for
  doing that)

Reviewers: csilvers, aasmund, benkraft

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D52344