Implement datastore writes and computed properties
Summary:
This commit adds the ability to do `DB.put()` and `DB.putAsync()` for our
data-class-based datastore models. The meat of this was doing the conversion
from the data classes to the Entity type. While implementing this, almost by
accident, I came upon a natural way to do the equivalent of ndb's computed
properties (which have the behavior of being stored in the datastore, but
ignored when being read back): a `val` defined on the data class that's not
part of the constructor. I went ahead and implemented these as well, since it
was easier than special casing not doing this!

I also went ahead and implemented the preferred version of `withMockDatastore`
that doesn't deal directly with `com.google.cloud.datastore.Entity`s and
instead uses our data classes.

Test Plan: `./gradlew test`

Reviewers: csilvers, aasmund, benkraft

Reviewed By: aasmund

Subscribers: samuel

Differential Revision: https://phabricator.khanacademy.org/D52426