Tweak the interface for constructing keys
Summary:

In D52343, @csilvers pointed out that the API for construcing keys was a bit
more complicated than it needs to be. I want to keep the `KeyIDOrName`
interface so that the type checker can help enforce that we've handled the
possibility of either correctly, but this commit adds some extra constructors
so that we can construct a key via `Key("SomeDatastoreKind", "a_key_name")` or
`Key("SomeDatastoreKind", 3L)`, for example.

I also made the parent path have a default value, since we expect the common
case to be that a key doesn't have a parent.

Test Plan:
- `./gradlew test`

Auditors: csilvers, aasmund, benkraft