Async usage and idiom/style fixes
Summary:
This is a follow-up to D52088 based on @aasmund's comments there. There's a few
idiom/style fixes: naming, comments, calling of lambdas-with-receiver, making
KeyPathElement a data class. The main fix, however, is to the initalization of
env and project, which I was doing totally incorrectly. This code had two
problems: I was not synchronizing everywhere I needed to, and also I was using
thread-based instead of coroutine-based synchronization, which can cause
deadlock. (In fact the only reason this wasn't deadlocking before was because I
was not synchronizing everywhere that was necessary...).

There's two things that needed to change: making the lazy intialization
of the datastore client coroutine-friendly, and making the DBEnvAndProject
synchronize using a coroutine-friendly mutex. In order to bridge properly to
blocking code, this required some additional wrapping up in `async` and
`runBlocking`. I also realized while doing more reading that we're going to
need to be careful to document that the initialization API we may be calling
`runBlocking`, as clients written in the single-threaded asynchronous style
with kotlin coroutines can be deadlocked when we do this. (Webapp is written
using java threads, rather than coroutines, so it's not really a concern
there.)  We mitigate this most places by running async datastore operations in
the IO threadpool (which grows to avoid deadlock), but this intialization code
can still pose a danger since it necessarily has to be called before we've set
up the DB coroutine context that uses this thread pool.

Test Plan:
- `./gradlew test`
- pull this new version into webapp
- `./gradlew test` there
- `./gradlew appengineRun` there, click around in the devserver, see
  conversions successfully logged

Reviewers: csilvers, aasmund, benkraft

Reviewed By: aasmund

Subscribers: samuel, aasmund

Differential Revision: https://phabricator.khanacademy.org/D52390