Implement queries (non-type-safe version)
Summary:
This commit implements an initial way to run datastore queries, using an
untyped map-like interface and corresponding DSL. (Eventually, we will also
provide an annotation-processor-based approach for generating type-safe query
builders for every model type, but that's more involved, and we'd like to
unblock the use of queries sooner.)

This was surprisingly difficult, not because of the actual
implementation (which though verbose, because the google datastore client
doesn't have any generic query tools and we therefore have to list out a lot of
things manually, was straightforward), but because of the test stub.

The core problem with the test stub is that the datastore client's public
interface includes the `Query` type, which is very opaque. We use the
`StructuredQuery` implemenation throughout the query implementation. All the
necessary methods to subclass this implementation, or to read out the filters
in the query are package-private, so there's not really any normal way for our
test stub to look at the query we've constructed and execute it.

This left us with two options:
- do something gross with reflection to read out the query filters
- also gross, define some code in the google cloud datastore package that acts
  as a shim to expose this information to the outside

I went with the latter approach of putting some shim code in the same package,
as it gives us more feedback about whether things will break at compile time,
rather than at run time if the private APIs change. The shim I decided to
create was one exposing the `.toPb()` methods on the various relevant
classes. The datastore protocol buffer types are public and have public
accessors from which we can read out the information we need, so this lets us
keep the surface area of the private APIs we're using to an absolute minimum.

Note also that the private APIs we're using are only for tests, so hopefully
there's very little production risk here as a result.

There was also one other unrelated difficulty that's worth mentioning: we'd
hoped to use a query builder DSL where we could use `==` and `<` and `>`,
etc. directly in the DSL. However, due to
https://youtrack.jetbrains.com/issue/KT-29316 we can't use `==` here, and
there's also no way to use the inequality operators because they're all
implemented in terms of a single function `compareTo`, which means that we have
no idea which operator was used in the DSL. I asked on the kotlin forums, and
it seems like the standard workaround is to defined infix functions called
`eq`, `lt`, `gt`, etc. that act as pseudo-operators. This is also consistent
with the naming for python's magic methods for these operations, so hopefully
this will not be too unfamiliar. So the (non-type-safe) DSL ends up looking
like:
```
DB.query<SomeModelClass> {
    "aProperty" eq 3
    "somethingElse" lt "abcd"
}
```
which is not quite as nice as we hoped, but is still (I think) an improvement
over manually constructing all the relevant filters. When we implement the
type-safe DSL via annotation processors, it should look just the same, except
that instead of strings on the left-hand-side of every operation, we have typed
properties.

Test Plan: - `./gradlew test`

Reviewers: csilvers, aasmund, benkraft

Reviewed By: csilvers

Subscribers: samuel

Differential Revision: https://phabricator.khanacademy.org/D52660