Bump the version number again
Summary:

Version 0.0.4 is now being used in production for webapp, so we want to make
sure it's not easy for someone to accidentally do a `./gradlew publish` and
have the next webapp deploy automatically pull in a version of code we didn't
intend or test thoroughly.

Eventually we'll want to handle this automatically as part of the release
process (INFRA-2851).

Test Plan:
- fingers crossed

Auditors: csilvers, aasmund, benkraft