Implement property indexing
Summary:
The java client library automatically indexes all properties, but at KA, we
generally default to not indexing properties unless we know we need to query
them. Thus, we need a mechanism for setting the indexing behavior.

This commit adds support for indexing via a `@Meta(indexed = true)` annotation,
and sets it up so that if this annotation is not present, we default to not
indexing anything. For our python code, we force every property to explicitly
set whether it's indexed or not, but that's sufficiently more verbose here
that I thought we might not want to have `@Meta(indexed = false)` everywhere.

In order to exclude properties from indexing, we have to use a slightly
different API for the datastore when setting properties (which uses the
interface `Value`, which is a wrapper of an actual value with metadata).

Test Plan: - `./gradlew test`

Reviewers: csilvers, aasmund, benkraft

Reviewed By: csilvers, aasmund

Subscribers: samuel

Differential Revision: https://phabricator.khanacademy.org/D52784