Implement default values for properties
Summary:
Previously, we didn't distinguish between the state where a value is missing in
the datastore, and where it's present but set to `null`. We still don't really
want to make this distinction in the public API. However, we do want to be able
to provide default values for properties (via defaults in the kotlin data class
constructor), and we don't want these to be substituted if we've actually
previously set a value explicitly to null in the datastore.

This commit implements default values and makes it so that we now distinguish
internally (via a sentinel value) between present but null and absent. We only
apply default values in the case where property is truly absent in the
datastore. This also hade the side-effect of turning a lot of null pointer
exceptions (though ones explicitly thrown at entity conversion time) into
illegal argument exceptions when calling the model constructor, which gives
nicer info about what was missing.

Note that this is an API breaking change, and this will require changes to
clients that may assume missing properties will be set automatically to null.

Test Plan: - `./gradlew test`

Reviewers: csilvers, aasmund, benkraft

Reviewed By: csilvers

Subscribers: samuel

Differential Revision: https://phabricator.khanacademy.org/D52785