Implement repeated properties (as List<T> in kotlin)
Summary:
This commit implements repeated values by conversion to/from kotlin's
lists. This matched up fairly well with the way the datastore library deals
with repeated properties, so this was fairly straightforward. The one edge case
to deal with is that the datastore client treats null and the empty list as the
same (always giving back the empty list, even if you set null), so I've just
handled this by only allowing non-nullable list types for repeated
properties. (Nullable elements are fine.)

Test Plan: - `./gradlew test`

Reviewers: csilvers, aasmund, benkraft

Reviewed By: csilvers

Subscribers: samuel

Differential Revision: https://phabricator.khanacademy.org/D52843