Add support for querying repeated fields.
Summary:
I think we just have to pass the query through to the underlying
client, who handles it fine; the problem, as always, is in our test
stub.  I had to reorganize the way we applied filters to make sure we
could apply the proper list-inequality rules.

Test Plan: ./gradlew test

Reviewers: colin

Reviewed By: colin

Subscribers: benkraft, samuel

Differential Revision: https://phabricator.khanacademy.org/D53086