Implement ndb-style StructuredProperty
Summary:
ndb has a concept of a `StructuredProperty`, which is like a
LocalStructuredProperty (entity-valued property) in interface, but flattens out
its values into (name-munged) top-level fields on an entity so that they can be
indexed and queried separately.

This is a bit of an ugly hack, but we need it for compatibility with python, so
this commit implements it. This got kind of messy and made even worse some of
the already bad code organization in EntityConversion.kt, largely because
unlike the rest of the property types, structured properties need access to the
values of multiple fields on the entities. As a result, we have to hook into
the conversion process relatively close to the API, and we end up having a very
different codepath for structured properties than for everything else. In
addition, repeated structured properties add extra confusion because they're
flattened out into several top-level repeated fields, so we have to do a lot of
extra work to take a List<StructuredProperty<T>> and extract all the individual
properties as lists.

Note that queries are not yet implemented; I'll be doing that in a follow-up
commit.

Test Plan:
- `./gradlew test`
- ran the following (in prod mode) to verify python compatibility:
```
data class StudentAssignmentData(
    val kaid: String
)

data class Assignment(
    val student_data: List<StructuredProperty<StudentAssignmentData>>,
    override val key: Key<Assignment>
) : Keyed<Assignment>

fun main() {
    val assignmentKey = Key<Assignment>("Assignment", 4503599640903680L)
    val result = DB.get(assignmentKey)
    println(result)
    println(result?.student_data?.map { it.value })
}
```
- verified the printed result was consistent with the data in the cloud console datastore viewer

Reviewers: csilvers, aasmund, benkraft

Reviewed By: csilvers

Subscribers: samuel

Differential Revision: https://phabricator.khanacademy.org/D53096