Allow querying by the kotlin name instead of just the datastore name
Summary:
When using datastore queries over properties with the `@Meta(name = ...)`
annotation, we found that we could only query using the name of the property in
the datastore, not the kotlin name. This commit changes it so that we can query
using either name.

The main change here is that we now need to pass along a class reference for
the model class being queried to the point where we're transforming the query
filters to their google datastore client representation, so that we can look up
names and see if there are any annotations for the given name. I also
refactored out some of the shared renaming code into its own file for better
reusability.

Test Plan: - `./gradlew test`

Reviewers: csilvers, aasmund, benkraft

Reviewed By: benkraft

Subscribers: samuel

Differential Revision: https://phabricator.khanacademy.org/D53213