Implement multi-get for datastore
Summary:
We will be implementing multi-get/put by fixed numbers of input/output.
That is to say that we will have getMulti take two abstract Keys and
return a Pair of TypedModels, take three and return a Triple, and so
forth. getMulti will be able to handle Keys of any types and return the
appropriate TypedModels.

This diff introduces get by 2, 3 and 4 Keys and its asynchronous counterpart.

Test Plan:
`./gradlew test`

Define getMulti for the mock datastore.
Test that we can get the expected pair of TypedModels given two Keys.
Test that we can handle the case where a Key returns null.

Reviewers: colin, aasmund, csilvers

Reviewed By: colin

Subscribers: benkraft

Differential Revision: https://phabricator.khanacademy.org/D53041