Implement GeoPt (Lat/Lng) properties
Summary:
The datastore supports location properties (called GeoPt in ndb and LatLng in
java) that consist of a location expressed in floating-point latitude and
longitude. Since this is a built-in datastore concept, this was fairly
straightforward: we just needed a kotlin representation and converters back and
forth to the google datastore library's version.

Test Plan: - `./gradlew test`

Reviewers: csilvers, aasmund, benkraft

Reviewed By: csilvers

Subscribers: samuel

Differential Revision: https://phabricator.khanacademy.org/D53601