Implement ndb-style date and time properties
Summary:
In addition to the standard datetime property, ndb supports both date and
time properties. These are implemented in ndb just by converting to datetime,
either at midnight on the given date, or at the given time on 1970-01-01.

This commit implements a similar thing for the kotlin datastore library,
where we now allow both LocalDate and LocalTime properties, and just convert
them to timestamps like for the existing LocalDateTime.

Test Plan: - `./gradlew test`

Reviewers: csilvers, aasmund, benkraft

Reviewed By: csilvers

Subscribers: samuel

Differential Revision: https://phabricator.khanacademy.org/D53610