Implement query ordering
Summary:
While we've already implemented basic filter-based datastore queries, we were
unable to order the result. This commit adds ordering, which is exposed in the
API as an extra optional argument to `DB.query`:
```
DB.query<T>("kind", orderBy("someField")) { }
```

Arguments are just passed along to the datastore client, and we rely on it to
do the actual ordering and/or check validity. The test stub implements the
sorting, but it does not attempt to validate anything (e.g. that indexes exist)
because the test stub doesn't have any notion of indexes.

Test Plan: - `./gradlew test`

Reviewers: csilvers, aasmund, benkraft

Reviewed By: csilvers

Subscribers: samuel

Differential Revision: https://phabricator.khanacademy.org/D53683