Implement JsonProperties
Summary:
ndb supports json properties, which store any json-serialized value as bytes in
the datastore. We've done monkeypatching so that these values are interpreted
as strings (primarily for the purposes of compatibility with bigquery), so this
commit implements JSON properties as a special property type that serializes to
strings.

Analogous to how we implemented structured properties, this is just implemented
as a wrapper around a value to which the JSON will be deserialized. In the
common case, this will probably be a data class, but it can also be a
`Map<String, String>` (or whatever we like) for dealing with schemaless JSON.

Test Plan: - `./gradlew test`

Reviewers: csilvers, aasmund, benkraft

Reviewed By: csilvers

Subscribers: samuel

Differential Revision: https://phabricator.khanacademy.org/D53731