Add a @Readonly annotation, for use with legacy models and projections
Summary:
We've got a whole bunch of legacy property types in our python code that it may
not make sense to translate to kotlin because we will never need to look at
them except from legacy python code.  Additionally, it's more efficient (and
cleaner code) to define model classes with only the subset of properties you
need to read (these function as a sort of projection in this library, though
the datastore client itself underneath is not using a projection, so we're
still sending all the data over the wire).

However, if we `.put()` a model without a complete set of properties in kotlin,
we will end up erasing all the other properties that are not included on the
model definition. This commit therefore adds an annotation `@Readonly` that we
check at `.put()` time, and if it's present, we'll throw instead of doing the
datastore write. This will give us more confidence in using incomplete kotlin
equivalents of our python models without worrying about data loss.

Test Plan: - `./gradlew test`

Reviewers: csilvers, aasmund, benkraft

Reviewed By: csilvers

Subscribers: samuel

Differential Revision: https://phabricator.khanacademy.org/D53770