Implement transactions in the test stub
Summary:
@kairui pointed out to me that the datastore test stub currently throws when
you try to use a transaction, which is making its usefulness limited. This
commit implementes very rudimentary transactions in the test stub. In
particular, this implementation does not correctly exclude concurrent
modification of an entity or an entity group by another
thread/coroutine. However, I expect this behavior to be rare in tests anyway,
so I left it as a TODO for now. I did implement the (potentially) most
confusing behavior of datastore transactions, which is that if you `.get()` an
entity you previously `.put()` in a transaction, you do not see the `.put()`
until after the transaction has committed.

Test Plan: - `./gradlew test`

Reviewers: csilvers, aasmund, benkraft

Reviewed By: csilvers

Subscribers: samuel, kairui

Differential Revision: https://phabricator.khanacademy.org/D53776