Version bump to 0.1.0, update README.
Summary:

Since we've finished all the functionality in the first phase, let's bump the
version to 0.1.0 to pull in the few miscelaneous things that aren't in a
release yet. I also updated the README to reflect current status.

I still left the version pre-1.0, since I think we'll probably want to do at
least a code organization and API polishing pass before we commit to not
breaking the public API.

Test Plan:
- fingers crossed

Auditors: csilvers, aasmund, benkraft