Allow queries with no filter conditions
Summary:
The function for converting datastore filters from our DSL to the google
cloud library's version throws if ther are no filters specified, because the
cloud library doesn't provide a constructor that's callable with no filters.
However, it is legal not to specify a filter at all on a query (at the level
of the query builder), which has the effect of returning all entities, but
our code was disallowing this.

This commit allows queries without filter conditions, which is implemented by
setting the filter on the query builder only if there are any filter
conditions.

Test Plan: - `./gradlew test`

Reviewers: csilvers, aasmund, benkraft

Reviewed By: csilvers

Subscribers: samuel

Differential Revision: https://phabricator.khanacademy.org/D54102