Give a more useful error message when failing to read a property
Summary:
If we fail to deserialize a property from the datastore, we get information
about, say, the cast that failed, but we don't get any information about
which property we were trying to deserialize at the time, which makes
debugging harder. This commit wraps all exceptions that are thrown when
deserializing properties in a custom exception type whose message indicates
the name of the property we were deserializing. By using the "cause"
constructor parameter for exceptions, we also maintain the original
exception's message and stacktrace.

Test Plan:
- `./gradlew test`
- rerun the tests in D54162, see that the failing ones now indicate the
  property name

Reviewers: csilvers, aasmund, benkraft

Reviewed By: csilvers

Subscribers: samuel

Differential Revision: https://phabricator.khanacademy.org/D54163