Add an interface for properties that do custom serialization / hooks
Summary:
Some users of the library have started doing writes to the datastore from
kotlin, but since we don't have our custom property types that flag data for
exporting to bigquery, this data isn't being exported. We need to be able to
implement those property types (not in this library, but in webapp), but we
can't do that easily because there's no existing hook in this library for
doing extra work when reading from / writing to the datastore.

This commit adds an interface for a custom property that can hook into the
property of serializing and deserializing to the datastore. This can be used
both for implementing custom serialization (we could, for instance, have a
property type that writes protobufs to the datastore), or for implementing
hooks for things like updating a timestamp on write.

Test Plan: - `./gradlew test`

Reviewers: csilvers, aasmund, benkraft

Reviewed By: csilvers

Subscribers: samuel

Differential Revision: https://phabricator.khanacademy.org/D54173