Simplify the interface for custom serialization properties
Summary:
In D54220, upon seeing the custom serialization property interface being
used, @csilvers commented that the separation between set/get and
toDatastore/fromDatastore was pretty confusing, and that it was difficult to
figure out what was being called when without referring back to the datastore
library code. This commit simplifies the interface down to what we discussed
there: just two methods, one called when reading from the datastore and
another called when writing to the datastore. (Also, we still require the
0-arg constructor.) This should hopefully make implementations of this
interface easier to understand.

Additionally, since the question of behavior when a property is missing in
the datastore came up, I added a test for this. It's no different from the
behavior of any other property (we take the default if present, and throw if
not), but it's still good to confirm explicitly.

Test Plan: - `./gradlew test`

Reviewers: csilvers, aasmund, benkraft

Reviewed By: csilvers

Subscribers: samuel, csilvers

Differential Revision: https://phabricator.khanacademy.org/D54261