Don't call `.setExcludeFromIndexes()` on list values
Summary:

When trying to write a list to the prod datastore, the google client library
threw an error saying that calling this function on a list was not allowed.
This commit removes the call. We still call it on the values inside the list,
which is the proper way to make them unindexed.

Test Plan:
- `./gradlew test`

Auditors: csilvers, aasmund, benkraft