Add additional tests for transaction context visibility
Summary:
See INFRA-3050. @kairui reported seeing a problem where a DB.get call
was behaving differently depending on whether it was directly inside a
DB.transactional or inside a function called from within a DB.transactional
block. Given how kotlin coroutine contexts work, this is a plausible way for
things to fail, so this commit adds some additional tests of this behavior.
These pass, so I've not yet succeeded at reproducing the bug, but these are
good to have regardless and give us a starting point for further
investigations.

Test Plan: - `./gradlew test`

Reviewers: csilvers, aasmund, benkraft

Reviewed By: csilvers

Subscribers: samuel, kairui

Differential Revision: https://phabricator.khanacademy.org/D54343