Implement queries in transactions
Summary:
See INFRA-3050. After some additional investigation @kairui determined that a
bug we were seeing where the test stub throws from a function called in a
transaction is actually just because we haven't implemented queries in
transactions in the test stub. This commit adds that functionality. This was
straightforward since we can just delegate to the non-transactional test stub
after doing a check that the query in question is an ancestor query.

Test Plan: - `./gradlew test`

Reviewers: csilvers, aasmund, benkraft

Reviewed By: csilvers

Subscribers: samuel, kairui

Differential Revision: https://phabricator.khanacademy.org/D54365