Use an InheritableThreadLocal to store the value of localDB
Summary:
@lizfaubell was seeing an issue where if you try to mock the datastore for a
dataflow pipeline using `withMockDatastore` this doesn't actually work
because the mocking only happens in the thread in which the test is running,
and the test dataflow pipeline runs in a separate thread pool that knows
nothing about the mock.

One solution here is to use an `InheritableThreadLocal` instead of a
`ThreadLocal` so that as long as you run the pipeline from within the mocking
context, the thread pool that gets started will inherit the mocked value.
However, there's a downside to this: if we were to start a thread from a
transaction, we'd inherit the transaction, and that could lead to undefined
behavior depending on whether the transaction commits before the background
thread uses it. Thus, I set it up so that we throw immediately if we try to
start a new thread in a transaction where this would be a problem.

Test Plan:
- `./gradlew test`
- pull this update into the problematic dataflow testing scenario, see that
  the tests now pass there.

Reviewers: csilvers, aasmund, benkraft, lizfaubell

Reviewed By: benkraft

Subscribers: lizfaubell

Differential Revision: https://phabricator.khanacademy.org/D55029