Adding getMulti and getMultiAsync for a list of keys
Summary:
Previously we had getMulti only for 2, 3, or 4 keys. This adds a function for getting a variable number of keys of the same type, passed in a list.

For discussion: we could think about adding some reasonable limits to the number of keys that can be passed in. Currently we rely on Google's limits: https://cloud.google.com/datastore/docs/concepts/limits I'm not sure what a reasonable limit would be for our use cases; it partly depends on the size of the data entries being returned, which can vary across type. We could also wait and see if we run into any issues in testing (and load testing?) and optimize then.

Test Plan: - ./gradlew test

Reviewers: colin, csilvers, aasmund

Reviewed By: colin

Subscribers: benkraft

Differential Revision: https://phabricator.khanacademy.org/D55096